requirements.txt
altgraph==0.17.3
Authlib==1.2.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.4
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
Flask-Login==0.6.2
fsspec==2023.9.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
pycparser==2.21
pyinstaller-hooks-contrib==2022.10
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sympy==1.12
termcolor==2.3.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
